backend/requirements.txt
# Core
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Server / static / DB
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==3.0.1
psycopg2-binary==2.9.10

# Common utils
requests==2.32.4
python-dotenv==1.1.1

# chatbot/RAG
openai==1.98.0
chromadb==1.0.15
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
tiktoken==0.9.0
